Improve state store API
